Move "avocado.virt" namespace to "avocado_virt"
Now, with the limitations for having a single namespace because
of the previous plugin architecture lifted, we can now have
custom namespaces for each module.

This is beneficial for various aspects, including that it shows
that there's no difference between the Avocado extensions led by
the project itself or completely external extensions.

The Avocado-Virt libraries and plugins now live on this new namespace,
avocado_virt, which will usually find it itself installed on
`/usr/lib/pythonX.Y/site-packages/avocado_virt`.
