+++ avocado_virt/defaults.py
@@ -19,7 +19,7 @@ Default values used in tests and plugin code.
 from avocado.core import data_dir
 from avocado.core.settings import settings
 from avocado.core.settings import SettingsError
-from avocado.virt.qemu import path
+from .qemu import path
 
 try:
     qemu_bin = settings.get_value('virt.qemu.paths', 'qemu_bin')
