+++ avocado_virt/plugins/virt.py
