+++ avocado_virt/plugins/virt_bootstrap.py
