+++ avocado_virt/qemu/machine.py
@@ -30,12 +30,12 @@ from avocado.utils import genio
 from avocado.utils import process
 from avocado.utils import wait
 from avocado.utils import path as utils_path
-from avocado.virt.qemu import monitor
-from avocado.virt.qemu import devices
-from avocado.virt.utils import image
+from . import monitor
+from . import devices
+from ..utils import image
 
 try:
-    from avocado.virt.utils import video
+    from ..utils import video
     VIDEO_ENCODING_SUPPORT = True
 except ImportError:
     VIDEO_ENCODING_SUPPORT = False
