+++ avocado_virt/qemu/monitor.py
