+++ avocado_virt/qemu/path.py
