+++ avocado_virt/test.py
@@ -17,7 +17,7 @@ import os
 
 from avocado import Test
 from avocado.utils import process
-from avocado.virt.qemu import machine
+from .qemu import machine
 
 
 class VirtTest(Test):
