+++ avocado_virt/utils/image.py
