+++ avocado_virt/utils/video.py
