+++ setup.py
@@ -52,10 +52,9 @@ setup(name='avocado-virt',
       author='Lucas Meneghel Rodrigues',
       author_email='lmr@redhat.com',
       url='http://github.com/avocado-framework/avocado-virt',
-      packages=['avocado',
-                'avocado.virt',
-                'avocado.virt.utils',
-                'avocado.virt.qemu',
-                'avocado.core.plugins'],
+      packages=['avocado_virt',
+                'avocado_virt.utils',
+                'avocado_virt.qemu',
+                'avocado_virt.plugins'],
       data_files=get_data_files()
       )
